fix: drop orphaned tool-result messages from coach history (PostHog #019d510a)
Gemini was rejecting message history with "function call turn must follow
user turn or function-response turn" — 158 occurrences across 57 users.
Failed/aborted streaming attempts left orphaned tool-result messages
(role: tool) in the agent component's persisted history. On retry the
shape became user → tool → user, which Gemini rejects.

stripOrphanedToolCalls now also drops tool role messages whose tool-result
parts reference no kept assistant tool-call. tool-approval-response parts
(keyed by approvalId, not toolCallId) are preserved.

Tests for stripOrphanedToolCalls move into their own file to keep both
under the 400-line cap.
